CLOUDP-110416: address PR comments
